+++ android/src/main/java/com/brentvatne/exoplayer/ExoPlayerView.java
@@ -1,18 +1,16 @@
 package com.brentvatne.exoplayer;
 
 import android.content.Context;
-
-import androidx.annotation.NonNull;
 import androidx.core.content.ContextCompat;
 import androidx.media3.common.AdViewProvider;
 import androidx.media3.common.C;
+import androidx.media3.common.Format;
 import androidx.media3.common.Player;
 import androidx.media3.common.Tracks;
 import androidx.media3.common.VideoSize;
 import androidx.media3.common.text.Cue;
 import androidx.media3.common.util.Assertions;
 import androidx.media3.exoplayer.ExoPlayer;
-import androidx.media3.exoplayer.trackselection.TrackSelectionArray;
 import androidx.media3.ui.SubtitleView;
 
 import android.util.AttributeSet;
@@ -26,6 +24,7 @@ import android.widget.FrameLayout;
 
 import com.brentvatne.common.api.ResizeMode;
 import com.brentvatne.common.api.SubtitleStyle;
+import com.google.common.collect.ImmutableList;
 
 import java.util.List;
 
@@ -37,8 +36,8 @@ public final class ExoPlayerView extends FrameLayout implements AdViewProvider {
     private final AspectRatioFrameLayout layout;
     private final ComponentListener componentListener;
     private ExoPlayer player;
-    final private Context context;
-    final private ViewGroup.LayoutParams layoutParams;
+    private Context context;
+    private ViewGroup.LayoutParams layoutParams;
     private final FrameLayout adOverlayFrameLayout;
 
     private boolean useTextureView = true;
@@ -112,7 +111,7 @@ public final class ExoPlayerView extends FrameLayout implements AdViewProvider {
     }
 
     public void setSubtitleStyle(SubtitleStyle style) {
-        // ensure we reset subtitle style before reapplying it
+        // ensure we reset subtile style before reapplying it
         subtitleLayout.setUserDefaultStyle();
         subtitleLayout.setUserDefaultTextSize();
 
@@ -210,6 +209,16 @@ public final class ExoPlayerView extends FrameLayout implements AdViewProvider {
         }
     }
 
+    /**
+     * Get the view onto which video is rendered. This is either a {@link SurfaceView} (default)
+     * or a {@link TextureView} if the {@code use_texture_view} view attribute has been set to true.
+     *
+     * @return either a {@link SurfaceView} or a {@link TextureView}.
+     */
+    public View getVideoSurfaceView() {
+        return surfaceView;
+    }
+
     public void setUseTextureView(boolean useTextureView) {
         if (useTextureView != this.useTextureView) {
             this.useTextureView = useTextureView;
@@ -236,40 +245,34 @@ public final class ExoPlayerView extends FrameLayout implements AdViewProvider {
         layout(getLeft(), getTop(), getRight(), getBottom());
     };
 
-    private void updateForCurrentTrackSelections() {
-        if (player == null) {
+    private void updateForCurrentTrackSelections(Tracks tracks) {
+        if (tracks == null) {
             return;
         }
-        TrackSelectionArray selections = player.getCurrentTrackSelections();
-        for (int i = 0; i < selections.length; i++) {
-            if (player.getRendererType(i) == C.TRACK_TYPE_VIDEO && selections.get(i) != null) {
-                // Video enabled so artwork must be hidden. If the shutter is closed, it will be opened in
-                // onRenderedFirstFrame().
+        ImmutableList<Tracks.Group> groups = tracks.getGroups();
+        for (Tracks.Group group: groups) {
+            if (group.getType() == C.TRACK_TYPE_VIDEO && group.length > 0) {
+                // get the first track of the group to identify aspect ratio
+                Format format = group.getTrackFormat(0);
+
+                // update aspect ratio !
+                layout.setAspectRatio(format.height == 0 ? 1 : (format.width * format.pixelWidthHeightRatio) / format.height);
                 return;
             }
         }
-        // Video disabled so the shutter must be closed.
+        // no video tracks, in that case refresh shutterView visibility
         shutterView.setVisibility(this.hideShutterView ? View.INVISIBLE : View.VISIBLE);
     }
-
-    Runnable hideShutterViewRunnable = new Runnable() {
-        @Override
-        public void run() {
-            shutterView.setVisibility(INVISIBLE);
-        }
-    };
 
     public void invalidateAspectRatio() {
         // Resetting aspect ratio will force layout refresh on next video size changed
         layout.invalidateAspectRatio();
-
-        removeCallbacks(hideShutterViewRunnable);
     }
 
     private final class ComponentListener implements Player.Listener {
 
         @Override
-        public void onCues(@NonNull List<Cue> cues) {
+        public void onCues(List<Cue> cues) {
             subtitleLayout.setCues(cues);
         }
 
@@ -286,18 +289,12 @@ public final class ExoPlayerView extends FrameLayout implements AdViewProvider {
 
         @Override
         public void onRenderedFirstFrame() {
-            // The shutter view is use to hide dirty resizing issue when starting playback.
-            // In case video doesn't match your View aspect ratio, at playback startup you may have flickering during resize.
-            // we saw that onRenderedFirstFrame is called before first resizing
-            // Then hiding the shutterView directly may not hide the flickering.
-            // This small delay avoid the flickering
-            postDelayed(hideShutterViewRunnable, 15);
+            shutterView.setVisibility(INVISIBLE);
         }
 
         @Override
-        public void onTracksChanged(@NonNull Tracks tracks) {
-            updateForCurrentTrackSelections();
+        public void onTracksChanged(Tracks tracks) {
+            updateForCurrentTrackSelections(tracks);
         }
     }
-
 }